FWWinのfullscreenモード対応にバグがいっぱいあったので，修正． src/tests/FWAppGLUTTestに，キー w を追加して， Full<->Winの切り替えに使えるようにした． あとは，warning 対応と DSTRの消し忘れの削除．
git-svn-id: http://springhead.info/spr2/Springhead2/trunk@3469 05cee5c3-a2e9-0310-9523-9dfc2f93dbe1

+++ src/Collision/CDCapsule.cpp
@@ -31,7 +31,7 @@ bool CDCapsule::IsInside(const Vec3f& p){
 }
 float CDCapsule::CalcVolume(){
 	return  4.0f/3.0f * M_PI * radius * radius * radius + 
-			M_PI * radius * radius * length;
+			(float)M_PI * radius * radius * length;
 }
 	
 // サポートポイントを求める
